Improve structure in plow cli
