+++ plow_cli/src/login.rs
@@ -1,4 +1,7 @@
+use crate::{config::create_configuration_directory_if_not_exists, feedback::command_not_complete};
 use anyhow::{anyhow, bail, Result};
+use clap::{arg, App, Command};
+use colored::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Debug, Deserialize)]
@@ -42,4 +45,25 @@ pub fn get_saved_api_token() -> Result<String> {
     let credentials_file_contents = std::fs::read_to_string(config_dir.join("credentials"))?;
     let credentials = toml::from_str::<CredentialsFile>(&credentials_file_contents)?;
     Ok(credentials.registry.token.to_owned())
+}
+
+pub fn attach_as_sub_command() -> App<'static> {
+    Command::new("login")
+        .about("Registers an api token for the CLI to use.")
+        .arg(arg!([API_TOKEN]))
+}
+
+pub fn run_command(sub_matches: &clap::ArgMatches) -> Result<()> {
+    create_configuration_directory_if_not_exists()?;
+    if let Some(token) = sub_matches.get_one::<String>("API_TOKEN") {
+        save_credentials_replace_existing(token)?;
+        println!(
+            "\t{} successful. Saved API token to ~/.plow/credentials.toml",
+            "Login".green().bold(),
+        );
+        return Ok(());
+    }
+    command_not_complete("please provide a valid API token to save");
+    // Unreachable code
+    Ok(())
 }
